core/requirements.txt
# SIGMAX Core Requirements

# ============================================
# LLM & Agents
# ============================================
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langgraph==0.2.59
langchain-openai==0.2.14
langchain-anthropic==0.3.3
langchain-groq==0.2.2
langchain-ollama==0.2.1
chromadb==0.5.23
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1

# ============================================
# Trading & Market Data
# ============================================
ccxt==4.4.41
freqtrade==2024.12
pandas==2.2.3
numpy==1.26.4
pandas-ta==0.3.14b
ta-lib==0.4.32
yfinance==0.2.50
alpaca-py==0.36.1
web3==7.6.0
solana==0.35.1
uniswap-python==0.7.7

# ============================================
# ML & RL
# ============================================
scikit-learn==1.6.1
stable-baselines3==2.4.0
gym==0.26.2
gymnasium==1.0.0
torch==2.5.1
tensorflow==2.18.0
keras==3.7.0
xgboost==2.1.3
lightgbm==4.5.0
catboost==1.2.7

# ============================================
# Quantum Computing
# ============================================
qiskit==1.3.1
qiskit-aer==0.15.1
qiskit-algorithms==0.3.1
qiskit-machine-learning==0.8.0
qiskit-optimization==0.6.1
qiskit-finance==0.4.1

# ============================================
# Portfolio & Risk
# ============================================
quantstats==0.0.62
pyportfolioopt==1.5.6
riskfolio-lib==6.3.1
empyrical==0.5.5

# ============================================
# API & Web
# ============================================
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
httpx==0.28.1
aiohttp==3.11.11
pydantic==2.10.4
pydantic-settings==2.7.0

# ============================================
# Database & Cache
# ============================================
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1
clickhouse-driver==0.2.9

# ============================================
# Message Queue
# ============================================
nats-py==2.9.1
aiokafka==0.12.0

# ============================================
# Security & Privacy
# ============================================
cryptography==44.0.0
pycryptodome==3.21.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================
# Telegram Bot
# ============================================
python-telegram-bot==21.9
aiogram==3.16.0

# ============================================
# Observability
# ============================================
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
sentry-sdk==2.19.2
loguru==0.7.3

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.1
pyyaml==6.0.2
toml==0.10.2
click==8.1.8
rich==13.9.4
tqdm==4.67.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pillow==11.0.0
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1

# ============================================
# Testing
# ============================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
locust==2.32.4

# ============================================
# Code Quality
# ============================================
black==24.10.0
ruff==0.8.5
mypy==1.14.0
pre-commit==4.0.1

# ============================================
# Development
# ============================================
ipython==8.31.0
jupyter==1.1.1
notebook==7.3.2
